refactor: domain 클래스들을 패키지로 묶음
