+++ src/main/java/blackjack/domain/card/Denominations.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Denominations {
     ACE("A", 1),
