+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Hand;
 
 import java.util.List;
