+++ src/main/java/blackjack/domain/participant/Players.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import blackjack.utils.CardDeck;
 import blackjack.utils.IllegalNameException;
