+++ src/main/java/blackjack/utils/CardDeck.java
@@ -1,6 +1,6 @@
 package blackjack.utils;
 
-import blackjack.domain.Card;
+import blackjack.domain.card.Card;
 
 import java.util.List;
