+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,9 +1,9 @@
-package blackjack;
+package blackjack.domain.participant;
 
-import blackjack.domain.Card;
-import blackjack.domain.Dealer;
-import blackjack.domain.Participant;
-import blackjack.domain.Player;
+import blackjack.domain.card.Card;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participant;
+import blackjack.domain.participant.Player;
 import blackjack.utils.CardDeck;
 import blackjack.utils.FixedCardDeck;
 import blackjack.utils.IllegalNameException;
